Start to migrate Amazon GuardDuty auditor
