+++ docs/setup/Setup_AWS.md
@@ -225,7 +225,7 @@ The ASM Module uses NMAP at its core and will be expanded to include ZAP and Sho
 
 ## AWS Checks & Services
 
-These are the following services and checks perform by each Auditor, there are currently **627 Checks** across **84 Auditors** that support the secure configuration of **120 services/components**
+These are the following services and checks perform by each Auditor, there are currently **627 Checks** across **85 Auditors** that support the secure configuration of **120 services/components**
 
 **Regarding AWS ElasticSearch Service/OpenSearch Service**: AWS has stopped supporting Elastic after Version 7.10 and released a new service named OpenSearch. The APIs/SDKs/CLI are interchangable. Only ASFF metadata has changed to reflect this, the Auditor Names, Check Names, and ASFF ID's have stayed the same.
 
@@ -423,6 +423,7 @@ These are the following services and checks perform by each Auditor, there are c
 | Amazon_EMR_Serverless_Auditor | EMR Serverless Application | Is Application in a VPC |
 | Amazon_EMR_Serverless_Auditor | EMR Serverless Application | Does Application use custom container runtime |
 | Amazon_EMR_Serverless_Auditor | EMR Serverless Application | Does Application configure auto-stopping jobs |
+| Amazon_GuardDuty_Auditor | GuardDuty detector | Is GuardDuty enabled |
 | Amazon_InspectorV2_Audtior | Inspector scanning configuration | Is Inspector V2 scanning enabled at all |
 | Amazon_InspectorV2_Audtior | Inspector scanning configuration | Is Inspector V2 scanning enabled for EC2 |
 | Amazon_InspectorV2_Audtior | Inspector scanning configuration | Is Inspector V2 scanning enabled for ECR |
@@ -682,7 +683,6 @@ These are the following services and checks perform by each Auditor, there are c
 | AWS_Secrets_Manager_Auditor | Secrets Manager secret | Is the secret over 90 days old |
 | AWS_Secrets_Manager_Auditor | Secrets Manager secret | Is secret auto-rotation enabled |
 | AWS_Security_Hub_Auditor | Security Hub hub | Is SecurityHub enabled |
-| AWS_Security_Services_Auditor | GuardDuty detector | Is GuardDuty enabled |
 | AWS_Security_Services_Auditor | Detective graph | Is Detective enabled |
 | AWS_Security_Services_Auditor | Macie2 session | Is Macie enabled |
 | ~~AWS_Security_Services_Auditor~~ | ~~AWS WAFv2 (Regional)~~ | ~~Are Regional Web ACLs configured~~ </br> **THIS FINDING HAS BEEN RETIRED** |
